+++ index.js
@@ -1,5 +1,5 @@
 var next = require('./lib/index.js');
 
-next.version = '1.14.4';
+next.version = '1.15.0';
 
 module.exports = next;