+++ src/field/index.js
@@ -29,6 +29,7 @@ class Field {
         this.fieldsMeta = {};
         this.cachedBind = {};
         this.instance = {};
+        this.initValue = options.values || {};
 
         this.options = Object.assign(
             {
@@ -94,10 +95,17 @@ class Field {
         )}`;
 
         const field = this._getInitMeta(name);
-        const defaultValue =
-            typeof initValue !== 'undefined'
-                ? initValue
-                : originalProps[defaultValueName];
+        let defaultValue;
+        if (typeof initValue !== 'undefined') {
+            defaultValue = initValue;
+        } else if (originalProps[defaultValueName]) {
+            defaultValue = originalProps[defaultValueName];
+        } else if (this.options.parseName) {
+            defaultValue = getIn(this.initValue, name);
+        } else {
+            defaultValue =
+                (this.initValue && this.initValue[name]) || undefined;
+        }
 
         Object.assign(field, {
             valueName,